chore: project refactoring
